backend/requirements.txt
fastapi
uvicorn
groq
firebase-admin
python-dotenv
pydantic